Switch to egit

+++ src/main/java/com/github/visola/githubnotifier/data/NotificationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 
 import com.github.visola.githubnotifier.model.Notification;
 
-public interface NotificationRepository extends PagingAndSortingRepository<Notification, Integer> {
+public interface NotificationRepository extends PagingAndSortingRepository<Notification, Long> {
 
 }
